+++ imagedephi/redact/__init__.py
@@ -1,3 +1,3 @@
-from .redact import ProfileChoice, iter_image_dirs, redact_images, show_redaction_plan
+from .redact import ProfileChoice, redact_images, show_redaction_plan
 
-__all__ = ["iter_image_dirs", "redact_images", "show_redaction_plan", "ProfileChoice"]
+__all__ = ["iter_image_dirs", "show_redaction_plan", "ProfileChoice"]
