+++ imagedephi/utils/constants.py
@@ -1,1 +1,2 @@
+IMAGE_DEPHI_MAX_IMAGE_PIXELS = 1000000000
 MAX_ASSOCIATED_IMAGE_SIZE = 160
